moving document filter to its own file
