+++ src/researchhub_document/models.py
@@ -1,8 +1,8 @@
 # flake8: noqa
+from .related_models.document_filter_model import DocumentFilter
 from .related_models.featured_content_model import FeaturedContent
 from .related_models.feed_exclusion_model import FeedExclusion
 from .related_models.researchhub_post_model import ResearchhubPost
 from .related_models.researchhub_unified_document_model import (
-    DocumentFilter,
     ResearchhubUnifiedDocument,
 )
